requirements.txt
arrow
python-dotenv
python-frontmatter
streamlit
sqlalchemy
sqlmodel
psycopg
psycopg-binary
langchain
langchain-community
langchain-openai
sentence-transformers
opensearch-py